Moved project to Use .Net 3.5 !!! More refactoring Fixed bug where DVD image was missing for DVDs with no Video TS folder Moved all Model items into their own folder
git-svn-id: https://videobrowser.googlecode.com/svn/trunk@779 096cc86f-8a4c-0410-9b85-8b5e0eee1c45

+++ MediaBrowser/Code/Application.cs
@@ -126,7 +126,7 @@ namespace MediaBrowser
                 object control = pi.GetValue(view, null);
 
                 pi = control.GetType().GetProperty("KeyRepeatThreshold", BindingFlags.NonPublic | BindingFlags.Instance);
-                pi.SetValue(control, val, null);
+                pi.SetValue(control, (UInt32)val, null);
             }
             catch
             {
